requirements.txt
discord.py
python-dotenv
pillow
wavesynth
requests
yt-dlp
aiosqlite
spotipy
